actualiacion

+++ Lector.java
@@ -2,21 +2,22 @@
 /**
  * Quien hace uso de los libros.
  * 
- * @author (your name) 
- * @version (a version number or a date)
+ * @author(Kevin Anaya, Laura Mendez)
+ * @version (1)
  */
 import java.time.LocalDate;
+import java.util.Scanner; 
 public class Lector extends Usuario
 {
     // instance variables - replace the example below with your own
     private LocalDate fechaNacimiento;
     private String nombre, apellidos, direccion;
-    private int telefono;
+    private long telefono;
 
     /**
      * Constructor for objects of class Lector
      */
-    public Lector(String nombreUsuario, String contraseña, String correo, int id, LocalDate fechaNacimiento, String nombre, String apellidos, String direccion, int telefono) {
+    public Lector(String nombreUsuario, String contraseña, String correo, int id, LocalDate fechaNacimiento, String nombre, String apellidos, String direccion, long telefono) {
         super(nombreUsuario, contraseña, correo, id);
         this.fechaNacimiento = fechaNacimiento;
         this.nombre = nombre;
@@ -57,11 +58,12 @@ public class Lector extends Usuario
         this.direccion = direccion;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
     public void setTelefono(int teelfono) {
         this.telefono = telefono;
     }
+
 }
